requirements.txt
aniso8601==3.0.2
autopep8==1.4.3
bandit==1.5.1
certifi==2018.11.29
coverage==4.5.2
Django==2.1.7
django-cors-headers==2.4.0
django-filter==2.1.0
djangorestframework==3.9.1
djangorestframework-jwt==1.11.0
entrypoints==0.3
flake8==3.7.6
flake8-import-order==0.18
flake8-pep257==1.0.5
gitdb2==2.0.5
GitPython==2.1.11
graphene==2.1.3
graphene-django==2.2.0
graphql-core==2.1
graphql-relay==0.4.5
mccabe==0.6.1
pbr==5.1.2
pep257==0.7.0
pep8==1.7.1
promise==2.2.1
psycopg2==2.7.6.1
psycopg2-binary==2.7.7
pycodestyle==2.5.0
pydocstyle==3.0.0
pyflakes==2.1.0
PyJWT==1.7.1
python-dotenv==0.10.1
pytz==2018.9
PyYAML==3.13
Rx==1.6.1
singledispatch==3.4.0.3
six==1.12.0
smmap2==2.0.5
snowballstemmer==1.2.1
stevedore==1.30.0
style==1.1.0
update==0.0.1
